Casos de diarréia em crianças

Estudo sobre o efeito da suplementação de vitamina A na diarreia no nordeste do Brasil em uma coorte de crianças de 6 a 48 meses são referentes ao ensaio comunitário randomizado placebo-controlado, realizado entre dezembro/1990 e dezembro/1991. Banco de dados disponível em Carvalho MS, Andreozzi VL, Codeço, C. T CDP, Barbosa MTS, Shimakura SE (2011). Análise de Sobrevivência: teoria e aplicações em saúde. Editora FIOCRUZ. e cedido pelo Instituto de Saúde Coletiva da Universidade Federal da Bahia. O banco de dados consiste em 860 crianças (426 no grupo placebo e 434 no grupo vitamina), gerando um total de 5.592 registros e 11 variáveis. Esse banco de dados é caracterizado pela presença de múltiplos eventos (cada criança pode ter um ou mais episódios de diarreia ao longo de sua permanência no estudo).

Format

Data frame com 5592 linhas e 11 variáveis:

numcri: identificador da criança

grupo: vit = receberam vitamina A, pla = placebo

sexo: fem = feminino, masc = masculino

idade: idade em meses no início do estudo

ini: data no início do período sem diarreia

fim: data no fim do período sem diarreia

diasant: duração em dias do episódio de diarreia anterior

mediadej: média de dejeções líquidas ou semilíquidas do episódio de diarreia anterior

enum: numeração das observações para cada criança

status: 0 = censura, 1 = evento (diarreia)

tempo: dias sem diarreia (fim-ini)

Source

http://sobrevida.fiocruz.br/diarreia.html

Carvalho MS, Andreozzi VL, Codeço, C. T CDP, Barbosa MTS, Shimakura SE (2011). Análise de Sobrevivência: teoria e aplicações em saúde. Editora FIOCRUZ.

Details

No arquivo, cada linha representa um tempo entre episódios de diarreia. As covariáveis constantes (grupo, sexo e idade) têm seus valores repetidos em todas as linhas referentes a essa criança. As covariáveis que mudam de episódio para episódio são diasant, mediadej e enum.

Author

Fábio N. Demarqui fndemarqui@est.ufmg.br