E1684 Melanoma Clinical Trial.

Two-arm phase III clinical trial conducted by the Eastern Cooperative Oncology Group (ECOG) to comparing high-dose interferon (IFN) to observation (OBS). There were a total of 286 patients diagnosed with melanoma enrolled in the study, which accrued from 1984 to 1990, and the study was unblinded in 1993.

Format

A data frame with 286 rows and 10 variables:

trt: treatment (1 - observation, 2 - IFN)

failtime: time to relapse from randomization

failcens: failure indicator (1 - failure; 0 - otherwise)

survtime: time from relapse to death

survcens: failure indicator (1 - failure; 0 - otherwise)

age: patient's age (in years)

node: 4 nodal categories

sex: patient's (0=male, 1=female)

ps: patient's performance status (0 - fully active; 1 - other)

breslow: Breslow depth (in mm)

References

Ibrahim JG, Chen MH, Sinha D (2001). Bayesian Survival Analysis, volume 1. Springer, New York, NY.

Author

Fabio N. Demarqui fndemarqui@est.ufmg.br