The Leukemia Survival Data in North-West England

A dataset on the survival of acute myeloid leukemia in 1,043 pateietns, first analyzed by (Henderson et al. 2002) . It is of interest to investigate possible spatial variation in survival after accounting for known subject-specific prognostic factors, which include age, sex, white blood cell count (wbc) at diagnosis, and the Townsend score (tpi) for which higher values indicates less affluent areas. Both exact residential locations of all patients and their administrative districts (24 districts that make up the whole region) are available.

A dataset on the survival of acute myeloid leukemia in 1,043 pateietns, first analyzed by (Henderson et al. 2002) . It is of interest to investigate possible spatial variation in survival after accounting for known subject-specific prognostic factors, which include age, sex, white blood cell count (wbc) at diagnosis, and the Townsend score (tpi) for which higher values indicates less affluent areas. Both exact residential locations of all patients and their administrative districts (24 districts that make up the whole region) are available.

Format

A sf/data frame with 1043 rows and 8 variables:

time: survival time in days

cens: right censoring status 0=censored, 1=dead

age: age in years

sex:male=1 female=0

wbc:white blood cell count at diagnosis, truncated at 500

tpi: the Townsend score for which higher values indicates less affluent areas

district:administrative district of residence

geometry: georeferenced data needed to plot maps.

Henderson R, Shimakura S, Gorst D (2002). “Modeling Spatial Variation in Leukemia Survival Data.” Journal of the American Statistical Association, 97(460), 965-972. doi:10.1198/016214502388618753 , https://doi.org/10.1198/016214502388618753.

A sf/data frame with 1043 rows and 8 variables:

time: survival time in days

cens: right censoring status 0=censored, 1=dead

age: age in years

sex:male=1 female=0

wbc:white blood cell count at diagnosis, truncated at 500

tpi: the Townsend score for which higher values indicates less affluent areas

district:administrative district of residence

geometry: georeferenced data needed to plot maps.

Henderson R, Shimakura S, Gorst D (2002). “Modeling Spatial Variation in Leukemia Survival Data.” Journal of the American Statistical Association, 97(460), 965-972. doi:10.1198/016214502388618753 , https://doi.org/10.1198/016214502388618753.

Author

Fábio N. Demarqui fndemarqui@est.ufmg.br