Ocorrências de doenças oportunistas em pessoas diagnosticadas com AIDS

Os dados utilizados para o tempo até a ocorrência de doenças oportunistas provêm de uma coorte de pacientes HIV positivos atendidos no Hospital Universitário Gaffrée e Guinle (Unirio). As variáveis foram obtidas a partir dos prontuários clínicos. Registrou-se, para cada paciente com Aids, o tempo até a ocorrência de cada doença ou sintomatologia caracteristicamente relacionada à imunodepressão: candidíase, tuberculose, alguns sinais hematológicos, herpes zoster, pneumonia por Pneumocistis.

Format

Data frame com 1195 linhas e 11 variáveis:

reg: identificador do paciente

sex: F = feminino, M = masculino

esc: 0 = sem escolaridade, 1 = até quatro anos de estudo, 2 = ensino fundamental, 3 = ensino médio, 4 = ensino superior

idade: idade em anos na entrada do estudo

udi: uso de drogas injetáveis: 0 = não, 1 = sim

sexual: comportamento sexual: 0 = não, 1 = sim

ini: data da entrada no estudo (diagnóstico de Aids)

oport: doença oportunista: Candida = candidíase, Tuberc = tuberculose, Hemato = algumas alterações hematológicas, Herpes = herpes zoster, Pneumo = pneumonia por Pneumocistis

fim: data do diagnóstico do evento relacionado à imunodepressão

status: 0 = censura, 1 = ocorrência do evento

tempo: fim - ini

Source

http://sobrevida.fiocruz.br/aidsoport.html

Carvalho MS, Andreozzi VL, Codeço, C. T CDP, Barbosa MTS, Shimakura SE (2011). Análise de Sobrevivência: teoria e aplicações em saúde. Editora FIOCRUZ.

Details

Os dados estão armazenados de maneira que cada linha do arquivo indica o tempo do diagnóstico de Aids até a ocorrência de uma doença oportunista.

Author

Fábio N. Demarqui fndemarqui@est.ufmg.br