Manutenção preventiva de transformadores

Uma companhia de energia elétrica deseja estabelecer uma política de manutenção preventiva para seus transformadores, visando minimizar o custo esperado associado a sua manutenção. Tal política de manutenção deve levar em conta ambos os custos decorrentes da paralização programada (manutenção preventiva) e não programada (falha e posterior reparo do tranformador). Com esse objetivo, 30 transformadores foram colocados sob teste e concluiu-se que o tempo ótimo entre manutenções preventivas é igual a 6285 horas.

Uma companhia de energia elétrica deseja estabelecer uma política de manutenção preventiva para seus transformadores, visando minimizar o custo esperado associado a sua manutenção. Tal política de manutenção deve levar em conta ambos os custos decorrentes da paralização programada (manutenção preventiva) e não programada (falha e posterior reparo do tranformador). Com esse objetivo, 30 transformadores foram colocados sob teste e concluiu-se que o tempo ótimo entre manutenções preventivas é igual a 6285 horas.

Format

Data frame com 61 linhas e 5 variáveis:

tranformador: indicadora do transformador sendo acompanhado

unidade: indicadora do tranformador sendo acompanhado (transformadores que passam por manutenção preventiva são considerados novas unidades)

preventiva: variável indicadora (1 - ocorrência de manutenção preventiva; 0 - caso contrário)

tempo: tempo até a falha do transformador (em horas)

evento: indicadora da ocorrência do evento de interesse (falha do tranformador)

Data frame com 61 linhas e 5 variáveis:

tranformador: indicadora do transformador sendo acompanhado

unidade: indicadora do tranformador sendo acompanhado (transformadores que passam por manutenção preventiva são considerados novas unidades)

preventiva: variável indicadora (1 - ocorrência de manutenção preventiva; 0 - caso contrário)

tempo: tempo até a falha do transformador (em horas)

evento: indicadora da ocorrência do evento de interesse (falha do tranformador)

References

Gilardoni GL, Colosimo EA (2007). “Optimal Maintenance Time for Repairable Systems.” Journal of Quality Technology, 39(1), 48-53. doi:10.1080/00224065.2007.11917672 , https://doi.org/10.1080/00224065.2007.11917672, https://doi.org/10.1080/00224065.2007.11917672.

Gilardoni GL, Colosimo EA (2007). “Optimal Maintenance Time for Repairable Systems.” Journal of Quality Technology, 39(1), 48-53. doi:10.1080/00224065.2007.11917672 , https://doi.org/10.1080/00224065.2007.11917672, https://doi.org/10.1080/00224065.2007.11917672.

Author

Fábio N. Demarqui fndemarqui@est.ufmg.br